Add vendor products to cart in mongo only until the api is ready
