+++ frontend/src/Services/ProductContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const ProductContext = createContext(null);
+
+export default ProductContext;
